+++ app/(root)/example-input/page.tsx
@@ -1,0 +1,5 @@
+import DemoInput from "@/components/module/example/demo-input";
+
+export default function Page() {
+  return <DemoInput />;
+}