+++ components/shared/layout/client.tsx
@@ -3,7 +3,6 @@ import { useEffect } from "react";
 import { usePathname } from "next/navigation";
 import useAuthStore from "@/store";
 import CoreDialog from "../dialog/core";
-import CoreDrawer from "../drawer/core";
 import { AppProgressBar as ProgressBar } from "next-nprogress-bar";
 import Loading from "../loading";
 
@@ -29,7 +28,6 @@ export default function Client({ children }: { children: React.ReactNode }) {
     <>
       {loading && <Loading isBg />}
       <CoreDialog />
-      <CoreDrawer />
       <ProgressBar
         height="4px"
         color="var(--primary)"