+++ lib/validation.ts
@@ -6,6 +6,30 @@ import {
   usernameRegex,
 } from "./regex";
 
+export const UserSchema = z.object({
+  email: z
+    .string({ required_error: "Email Harus diisi" })
+    .email()
+    .refine(() => true, { message: "Email Harus diisi" }),
+  password: z
+    .string({ required_error: "Password Harus diisi" })
+    .min(6, "Password minimal 6 karakter")
+    .max(50, "Password maksimal 50 karakter"),
+  name: z
+    .string({ required_error: "Nama Harus diisi" })
+    .min(5, nameRegex.msg.min)
+    .max(50, nameRegex.msg.max)
+    .regex(nameRegex.regex, "Nama harus berupa huruf"),
+  username: z
+    .string({ required_error: "Username Harus diisi" })
+    .min(5, nameRegex.msg.min)
+    .max(15, "Maximum 15 characters")
+    .regex(usernameRegex.regex, usernameRegex.msg),
+  hobby: z
+    .string({ required_error: "Hobby Harus diisi" })
+    .regex(nameRegex.regex, "Hobby harus berupa huruf"),
+});
+
 export const ProfileSchema = z.object({
   email: z
     .string({ required_error: "Email Harus diisi" })